backend_api/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0

SQLAlchemy==2.0.44
asyncpg==0.28.0
alembic==1.11.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

pydantic>=2.3
pydantic-settings>=2.2

python-multipart==0.0.6
python-dotenv==1.0.0

greenlet==3.0.3

email-validator>=2.0.0